Add revenue module for tariff design (#661)
* Create Tariff Design.md

* Update Tariff Design.md

* Update Tariff Design.md

* Update Tariff Design.md

* Update Tariff Design.md

* Update Tariff Design.md

* Update Tariff Design.md

* PGandE Residential 

Updated tariffs

* Stop time change for verification

* Add support for parent class definition in loader

* Add support for more general tariff design specification

* Update Tariff.md

* Update pgande_res_200301.glm

* Add SDG&E residential tariffs (#797)

* SDG&E res rates

DR rates Effective rates for June 1, 2020
TOU DR rates effective rates for April 1, 2020
Low income discount and climate credits to be revised

* PG&E rates October 1, 2020

Tariffs changes in PG&E Residential rates

* Restoring values

Restoring values for tariffs dated March 1, 2020

* New name for SDGandE

Name change

* Delete SDG&E_res_200601.glm

* PCIA rates pgande

* value check

* script check

* Testing

* Adding revenue module

* completion check

* billing module inclusion

* changes

* revert module change

* structure check

* sce residential tariffs

* change baseline

* minor changes

* folder inclussion

* update extension


* Cleanup of revenue models and support data

* Documentation

* Rename SDGandE_res_200601.glm to sdgande_res_200601.glm

* Fix autotest failure

* Create billing.csv

* Update to PG&E revenue autotest

* Fix bill day in revenue autotest

* Try again to fix billing autotest

* Still trying to fix it

* Update test_pgande_billing.glm

* Fix for ABOVE_NORMAL and NORMAL Rfloor value

* Update test_geodata.txt
