+++ docs/Use cases/Tariff Design.md
@@ -1,0 +1,252 @@
+# Tariff Design [TO DO]
+
+This use-case is designed to include a number of tariff structures that IOUs, 
+POUs and CCAs implement to give the user the opportunity to explore a variety of tariff designed for the network in quesiton. 
+Below the tariff structures are outlined.
+
+## Investor Owned Utilities 
+
+### Bear Valley Electric Service
+
+#### General (Commercial)
+The general service has a break down of Small (<20 kW monthly maximum demand ) / Medium (20kW - 50 kW  monthly maximum demand) / Large (50kW - 200kW monthly maximum demand)
+Note, the values below exemplify Small general service to emphasize the structure of the tariff design. Medium and large prices and thresholds will vary.
+- Service Charges: $0.450 per meter per day 
+- Demand Charges: $0 
+      - Summer and winter breakdown ($ per kW per meter per month)
+- Energy Charges Summer / Winter breakdown ($/kWh): 2-tier system with 49.3 kWh day threshold  
+  - tier 1: 0.19145
+  - tier 2: 0.22582
+- Other Energy Charges ($/kWh):
+  - PPPC 6: 0.00881
+  - Taxes and fees7: 0.00088
+  - MHP BTM Capital Project: 0.00194
+  - Fire Hazard Protection: 0.00210
+  - Renewable Portfolio Standard: 0.00322
+- Minimum Charge: equal to the Service Charge per meter, per day.
+
+
+[Ref](https://www.bves.com/media/managed/adviceletter373and375/ELEC_SCHEDS_No_A_1_General_Service.pdf)
+
+#### Time of Use (200kW - 500 kW maximum monthly demand)
+
+- Service Charge per meter, per day: $16.40
+- Demand Charges: 
+      - Monthly non TOU Maximum Demand ($ per kW per meter per month)
+      - Summer month on-peak supply ($ per kW per meter per month)
+      - Summer month on-peak base ($ per kW per meter per month)
+      - Winter month on-peak supply ($ per kW per meter per month)
+      - Winter month on-peak base ($ per kW per meter per month)
+ - Energy Charges ($/kWh)
+      - Summer and Winter break down
+      - On-peak: 0.18542
+      - Mid-peak: 0.16303
+      - Off-peak: 0.14810
+ - Other Energy Charges ($/kWh):
+  - PPPC 6: 0.00881
+  - Taxes and fees7: 0.00088
+  - MHP BTM Capital Project: 0.00194
+  - Fire Hazard Protection: 0.00210
+  - Renewable Portfolio Standard: 0.00322
+- Minimum Charge: equal to the Service Charge per meter, per day + $3.00 per kW multuplied by the Contract Demand
+
+#### Residential 
+
+Total energy rates ($/kWh)
+- Baseline usage: $0.24373
+- 101%-400% Baseline: $0.30672
+- High Usage over 400% baseline: $0.53738
+Delivery Minimum bill amount ($ per meter per day): $0.32854
+California Climate Credit (per household, per semi-annual payment occuring in the April and October bill cycles): $35.73
+
+### Liberty Utilities (a.k.a CalPeco for California Electric Co.)
+
+
+
+### Pacific Gas and Electric (PG&E)
+[Ref](https://www.pge.com/tariffs/index.page)
+#### General
+
+Total Customer Charge Rates
+- Customer Charge Single-phase ($ per meter per day): $0.32854
+- Customer Charge Poly-phase ($ per meter per day): $0.82136
+Total Energy Rates ($/kWh)
+- Summer: $0.28091
+- Winter: $0.22036
+
+#### TOU
+Total Customer Charge Rates
+- Customer Charge Single-phase ($ per meter per day: $0.32854
+- Customer Charge Poly-phase ($ per meter per day): $0.82136 
+Total TOU Energy Rates ($ per kWh) 
+- Peak Summer: $0.29592
+- Part-Peak Summer: $0.27227
+- Off-Peak Summer: $0.24491
+- Part-Peak Winter $0.25166
+- Off-Peak Winter: $0.23075
+PDP Rates (Consecutive Day and Four-Hour Event Option:
+- PDP Charges ($ per kWh)
+      - All Usage During PDP Event: $0.60
+- PDP Credits Energy ($ per kWh) 
+      - Peak Summer: $0.00905
+      - Part-Peak Summer: $0.00905
+      - Off-Peak Summer: $0.00905
+      
+      
+[Small General A-1](https://www.pge.com/tariffs/assets/pdf/tariffbook/ELEC_SCHEDS_A-1.pdf)
+
+### PacifiCorp
+
+### San Diego Gas and Electric 
+
+### Southern California Edison (SCE)
+
+## Publicly Owned LSEs Including Publicly Owned Utilities (POUs)
+
+### Alameda Municipal Power
+
+### City of Anaheim
+
+### Azusa Light and Water
+
+### City of Banning
+
+### Biggs Municipal Utilities
+
+### Burbank Water and Power
+
+### CCSF (also called the Power Enterprise of the San Francisco Public Utilities Commission)
+
+### City of Cerritos, Cerritos Electric Utility
+
+### City of Industry
+
+### Colton Public Utilities
+
+### City of Corona
+
+### Eastside Power Authority
+
+### Glendale Water and Power
+
+### Gridley Electric Utility
+
+### City of Healdsburg
+
+### Imperial Irrigation District (IID)
+
+### Kirkwood Meadows Public Utility District
+
+### Lassen Municipal Utility District
+
+### Lathrop Irrigation District
+
+### Lodi Electric Utility
+
+### City of Lompoc
+
+### Los Angeles Department of Water & Power (LADWP)
+
+### Merced Irrigation District (MeID)
+
+### Modesto Irrigation District (MID)
+
+### Moreno Valley Utility (MVU)
+
+### City of Needles (Public Utility Authority)
+
+### City of Palo Alto
+
+### Pasadena Water and Power
+
+### City of Pittsburg, Pittsburg Power Company Island Energy
+
+### Port of Oakland
+
+### Port of Stockton
+
+### Power and Water Resources Pooling Authority (PWRPA)
+
+### Rancho Cucamonga Municipal Utility
+
+### Redding Electric Utility
+
+### City of Riverside
+
+### Roseville Electric
+
+### Sacramento Municipal Utility District (SMUD) 
+
+### City of Shasta Lake 
+
+### Shelter Cove Resort Improvement District
+
+### Silicon Valley Power (SVP)
+
+### Trinity Public Utilities District (PUD)
+
+### Truckee Donner Public Utilities District
+
+### Turlock Irrigation District (TID)
+
+### City of Ukiah
+
+### City of Vernon
+
+### Victorville Municipal Utilities Services
+
+## Rural Electric Cooperatives
+
+### Anza Electric Cooperative, Inc.
+
+### Plumas-Sierra Rural Electric Cooperative
+
+### Surprise Valley Electrification Corporation
+
+### Valley Electric Association, Inc.
+
+## Community Choice Aggregators 
+### Apple Valley Choice Energy
+
+### Clean Power SF
+
+### Clean Power Alliance
+
+### Desert Community Energy
+
+### Lancaster Choice Energy
+
+### Lancaster Power Authority
+
+### Marin Clean Energy (MCE)
+
+### Metropolitan Water District of So. Cal. (MWD)
+
+### Monterey Bay Community Power
+
+### Peninsula Clean Energy
+
+### Pico Rivera Innovative Municipal Energy
+
+### Pilot Power Group, Inc.
+
+### Pioneer Community Energy
+
+### Rancho Mirage Energy Authority
+
+### Redwood Coast Energy Authority
+
+### San Jacinto Power
+
+### San Jose Clean Energy
+
+### Silicon Valley Clean Energy
+
+### Solana Energy Alliance
+
+###  Clean Power
+
+### Valley Clean Energy
+
+
+*The list of utilities in California is obtained from [ca.gov](https://ww2.energy.ca.gov/almanac/electricity_data/utilities.html)
