+++ module/revenue/autotest/billing.csv
@@ -1,0 +1,2 @@
+datetime,meter,tariff,billing_days,energy,demand,charges
+2020-01-26,meter_1,E1,25,388.2,0,102.45
