+++ module/revenue/autotest/pgande_pcia_vintage.csv
@@ -1,0 +1,1 @@
+2009,0.025612010,0.028772011,0.031222012,0.031092013,0.033252014,0.033722015,0.033542016,0.03342017,0.033342018,0.033362019,0.035552020,0.04243
