+++ module/revenue/autotest/pgande_res_201001.glm
@@ -1,0 +1,64 @@
+module revenue
+{
+//	lowincome_discount -34.682 %;
+//	program_credit 17.87 $;
+//	program_credit_months [4,10];
+//	summer_season_months [6,7,8,9];
+//	winter_season_months [1,2,3,4,5,10,11,12];
+	billing_module "test_billing";
+	billing_library "${TESTDIR:-.}";
+};
+
+object tariff {
+	name "E1";
+	rate_design "Tiered Energy Charges";
+	minimum_daily_charge 0.32854 $/day;
+	energy_charge_base 0.24373 $/kWh;
+	energy_charge_100 0.30672 $/kWh;
+	energy_charge_400 0.53738 $/kWh;
+	average_total_rate 0.26249 $/kWh;
+}
+
+object tariff {
+	name "EM";
+	rate_design "Tiered Energy Charges";
+	minimum_daily_charge 0.32854 $/day;
+	energy_charge_base 0.24373 $/kWh;
+	energy_charge_100 0.30672 $/kWh;
+	energy_charge_400 0.53738 $/kWh;
+	average_total_rate 0.26249 $/kWh;
+}
+
+object tariff {
+	name "ES";
+	rate_design "Tiered Energy Charges";
+	minimum_daily_charge 0.32854 $/day;
+	discount 0.03115 $;
+	minimum_average_rate_limit 0.04892 $/kWh;
+	energy_charge_base 0.24373 $/kWh;
+	energy_charge_100 0.30672 $/kWh;
+	energy_charge_400 0.53738 $/kWh;
+	average_total_rate 0.26249 $/kWh;
+}
+
+object tariff {
+	name "ESR";
+	rate_design "Tiered Energy Charges";
+	minimum_daily_charge 0.32854 $/day;
+	energy_charge_base 0.24373 $/kWh;
+	energy_charge_100 0.30672 $/kWh;
+	energy_charge_400 0.53738 $/kWh;
+	average_total_rate 0.26249 $/kWh;
+}
+
+object tariff {
+	name "ET";
+	rate_design "Tiered Energy Charges";
+	minimum_daily_charge 0.32854 $/day;
+	discount 0.06181 $;
+	minimum_average_rate_limit 0.04892 $/kWh;
+	energy_charge_base 0.24373 $/kWh;
+	energy_charge_100 0.30672 $/kWh;
+	energy_charge_400 0.53738 $/kWh;
+	average_total_rate 0.26249 $/kWh;
+}
