+++ module/revenue/autotest/sdgande/sdgande_res_201001.glm
@@ -1,0 +1,30 @@
+//
+module revenue
+{
+	// lowincome_discount  %;
+	// program_credit  $;
+	// program_credit_months [];
+	// summer_season_months [6,7,8,9,10];
+	// winter_season_months [1,2,3,4,5,11,12];
+	billing_module "test_billing";
+	billing_library "${TESTDIR:-.}";
+};
+
+class sce_residential_tariff{
+	parent tariff;
+	double minimum_daily_charge[$/day];
+	double energy_charge_base[$/kWh];
+	double energy_charge_130[$/kWh];
+	double energy_charge_400[$/kWh];
+	char256 billing_function;
+
+object sce_residential_tariff {
+	name "DR";
+	rate_design "Tiered Energy Charges";
+	minimum_daily_charge 0.338 $/day;
+	energy_charge_base 0.27539 $/kWh;
+	energy_charge_130 0.34675 $/kWh;
+	energy_charge_400 0.43344 $/kWh;
+
+}
+
