+++ module/revenue/autotest/test_pgande_billing_bill_1.csv
@@ -1,0 +1,3 @@
+timestamp,billing_days,energy_charges,total_charges,total_bill
+2020-01-01T00:00:00-08:00,0,+0,+0,+0
+2020-01-26T01:00:00-08:00,25,+102.449,+102.449,+102.449
